+++ spec/client_spec.rb
@@ -1,7 +1,7 @@
 # encoding: UTF-8
 require_relative 'spec_helper'
 
-describe 'openstack-metering::client' do
+describe 'openstack-telemetry::client' do
 
   describe 'ubuntu' do